Fix sendall and recvall

+++ network.h
@@ -2,6 +2,6 @@
 
 int server(char *port);
 int client(char *host, char *port);
-int sendall(int sock, void *buf, size_t size);
-int recvall(int sock, void *buf, size_t size);
+ssize_t sendall(int sock, void *buf, ssize_t size);
+ssize_t recvall(int sock, void *buf, ssize_t size);
 char *addrstr(struct sockaddr *sa);
